backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
dj-database-url==2.1.0
python-dotenv==1.0.1
whitenoise==6.6.0
pandas==2.2.3
openpyxl==3.1.2
ofxparse==0.21
python-dateutil==2.9.0
